html is fully compiled, the header section, the styles of the first and second sections are given in full
the html codes of the page were written in full. The general css styles and header section styles of the page, the styles of the first and second sections of the main part, are fully subdivided. an attempt was made to give the icons used in the second section with a pseudo-element